// frontend/src/api/generated/model/scoutStartupMatchesResponseDtoDataItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export type ScoutStartupMatchesResponseDtoDataItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  investorId: string;
  /** @nullable */
  investorName?: string | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  thesisFitScore?: number | null;
  /** @nullable */
  fitRationale?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  createdAt?: string | null;
};